[REFACTOR][E] Rename package "saros.ui.widgets.viewer.project"
